requirements.txt
pandas
numpy
sentence-transformers
umap-learn
scikit-learn
plotly
streamlit